+++ CHANGELOG.md
@@ -70,6 +70,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added ability to configure electronOptions
 - Calendar module: option to hide private events
 - Add root_path for global vars
+- Added console warning on startup when deprecated config options are used 
 
 ### Updated
 - Modified translations for Frysk.